requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
websockets==12.0
python-multipart==0.0.6

# Trading
ccxt==4.1.0
pandas==2.1.3
numpy==1.26.2
ta==0.10.2

# Database
PyMySQL==1.1.0
sqlalchemy==2.0.23
redis==5.0.1

# Utils
python-dotenv==1.0.0
pydantic==2.5.0
structlog==23.2.0
python-telegram-bot==20.7
APScheduler==3.10.4

# ML (optional for now)
scikit-learn==1.3.2
xgboost==2.0.2